make do_populate_lic MACHINE specific
The license directory was made machine specific since bitbake
will otherwise error that the license would be overwritten by
the version of the other machine. However qa now kicks in and
warns that the license for non MACHINE specific packages are
missing. This make the do_populate lic MACHINE specific to
prevent that.