mesa: Upgrade to 19.3.1
Upgrade mesa to 19.3 and refresh the patches.

As MESA_EGL_NO_X11_HEADERS was renamed to EGL_NO_X11 in [1] we are going
to change the do_install and add a patch to libepoxy to use the new
define.

1: https://gitlab.freedesktop.org/mesa/mesa/commit/6202a13b71e18dc31ba7e2f4ea915b67eacc1ddb

Signed-off-by: Alistair Francis <alistair@alistair23.me>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
(cherry picked from commit 7a3d575ba56311c46122dfe9fa88094859fe9713)